added type annotation for some variables
This is to help further increase readability

+++ app/products.py
@@ -1,10 +1,11 @@
+import MySQLdb
 from flask import Blueprint, request, jsonify, Response
 
 product_bp = Blueprint('product', __name__, url_prefix='/')
 
-
 no_key_found_message = "no key found"
 bad_request_message = "bad request"
+
 
 def import_query_run_function():
     from run_query import run_sql_script
@@ -18,13 +19,12 @@ Get all the products in the database
 
 @product_bp.route("/product/all", methods=['GET'])
 def all_products() -> Response:
-
     query_run_function = import_query_run_function()
 
     output_rows, mysql, cur = query_run_function(f"SELECT * FROM products")
 
     if output_rows > 0:
-        products = cur.fetchall()
+        products: MySQLdb.cursors.CursorStoreResultMixIn = cur.fetchall()
         cur.close()
         return jsonify(products)
     return jsonify(success=False, message=no_key_found_message)
@@ -39,12 +39,12 @@ Get a certain product from the datbase provided with the id
 def product() -> Response:
     query_run_function = import_query_run_function()
     if request.method == 'GET':
-        product_id = request.args.get('id')
+        product_id: str = request.args.get('id', type=str)
 
         output_rows, mysql, cur = query_run_function(f"SELECT * FROM products WHERE product_id={product_id}")
 
         if output_rows > 0:
-            pdt = cur.fetchone()
+            pdt: MySQLdb.cursors.CursorStoreResultMixIn = cur.fetchone()
             return jsonify(pdt)
         return jsonify(success=False, message=no_key_found_message)
     else:
@@ -56,19 +56,19 @@ Delete a product from the database
 '''
 
 
-@product_bp.route('/product/delete')
+@product_bp.route('/product/delete', methods=["POST"])
 def delete_product() -> Response:
     query_run_function = import_query_run_function()
-    try:
-        product_id = request.args.get('id')
+    if request.method == "POST":
+        product_id: str = request.args.get('id', type=str)
 
         output_rows, mysql, cur = query_run_function(f"DELETE FROM products WHERE product_id = {product_id}")
 
         mysql.connection.commit()
         cur.close()
-
+        request.method = 'GET'
         return all_products()
-    except:
+    else:
         return jsonify(success=False, message=bad_request_message)
 
 
@@ -83,14 +83,15 @@ def create_product() -> Response:
     query_run_function = import_query_run_function()
 
     if request.method == 'POST':
-        name = request.args.get('name')
-        price = request.args.get('price')
+        name: str = request.args.get('name', type=str)
+        price: str = request.args.get('price', type=str)
 
-        output_rows, mysql, cur = query_run_function(f"INSERT INTO products(product_name, price) VALUES('{name}',{price})")
+        output_rows, mysql, cur = query_run_function(
+            f"INSERT INTO products(product_name, price) VALUES('{name}',{price})")
 
         mysql.connection.commit()
         cur.close()
-
+        request.method = 'GET'
         return all_products()
     else:
         return jsonify(success=False, message=bad_request_message)
@@ -105,11 +106,10 @@ Lists all the listings in the database in JSON format
 def edit_product() -> Response:
     query_run_function = import_query_run_function()
     if request.method == 'POST':
-        product_id = request.args.get('id')
-        name = request.args.get('name')
-        price = request.args.get('price')
-        print(product_id, name, price)
-        (output, mysql, cur) = query_run_function(
+        product_id: str = request.args.get('id', type=str)
+        name: str = request.args.get('name', type=str)
+        price: str = request.args.get('price', type=str)
+        output, mysql, cur = query_run_function(
             f"UPDATE products SET product_name = '{name}', price = {price} WHERE product_id={product_id}")
 
         mysql.connection.commit()
